fix code smells for sonarcloud

+++ net/websocket/include/WebSocketClient.hpp
@@ -36,7 +36,7 @@ public:
 
     long sendBinaryMessage(ByteArray data);
 
-    int sendFile(File);
+    long sendFile(File);
 
     void close();
